move send selectors to selectors dir
