+++ ui/app/pages/send/send-content/add-recipient/ens-input.container.js
@@ -3,7 +3,7 @@ import {
   getCurrentNetwork,
   getSendTo,
   getSendToNickname,
-} from '../../send.selectors'
+} from '../../../../selectors/send'
 import {
   getAddressBookEntry,
 } from '../../../../selectors'
