+++ ui/app/pages/send/send-content/send-row-wrapper/send-row-error-message/send-row-error-message.container.js
@@ -1,5 +1,5 @@
 import { connect } from 'react-redux'
-import { getSendErrors } from '../../../send.selectors'
+import { getSendErrors } from '../../../../../selectors/send'
 import SendRowErrorMessage from './send-row-error-message.component'
 
 export default connect(mapStateToProps)(SendRowErrorMessage)
