+++ ui/app/pages/settings/contact-list-tab/add-contact/add-contact.container.js
@@ -5,7 +5,7 @@ import { withRouter } from 'react-router-dom'
 import { addToAddressBook, showQrScanner, qrCodeDetected } from '../../../../store/actions'
 import {
   getQrCodeData,
-} from '../../../send/send.selectors'
+} from '../../../../selectors/send'
 
 const mapStateToProps = (state) => {
   return {
